feat(dataset): add dataset related interfaces
Add `Dataset` and `Product` interfaces that abstract the dataset and
dataset's product, respectively. The `Dataset` interface provides
methods to get the location of products within the dataset folder
structure and the requested date ranges. The `Product` interface
provides methods to match the product's filename with the dataset's
product pattern and to extract the date from the filename.