refactor(downloader): update get_files return type to include file paths
Update the return type of the `get_files` method in the `Downloader` class to include both the file paths and the file objects. This change is made to provide more information about the files retrieved from the datasource. The return type is now a list of tuples, where each tuple contains the file path and the corresponding file object.

Refactor the `get_files` method to retrieve the file paths using the `get_file_list` method and then retrieve the file objects using the `retrieve_files` method. The retrieved file objects are then zipped with their respective file paths and returned as a list of tuples.

This change improves the clarity and usability of the `get_files` method in the `Downloader` class.